Add validator to DrugType Edit/Create

+++ web/SU/drugtypes.jsp
@@ -20,12 +20,11 @@
                     <td><s:url id="deleteURL" action="deletedrugtype">
                         <s:param name="id" value="%{id}"></s:param>
                         </s:url>
-                        <s:a href="%{deleteURL}">Delete</s:a></td>
+                        <s:a href="%{deleteURL}" onclick="return confirm('Are you sure?')">Delete</s:a></td>
                     <td><s:url id="editURL" action="editdrugtype">
                         <s:param name="id" value="%{id}"></s:param>
                     </s:url>
                         <s:a href="%{editURL}">Edit</s:a></td>
-                    <td><a href="#">Info</a></td>
                 </tr>
             </s:iterator>
         </table>
